Move Symbol out of lib.rs
